requirements/base.txt
Django==2.0.4
python-decouple==3.1
pytz==2018.4
Unipath==1.1
